Update 10/31
Made command decks pretty dynamic.  Not fully, it's just able to make an infinitely long command deck without any info as to what those commands do, but it's good progress and always keeps track of the back button.  Maybe worth seeing if I can change the button from index to tracking the text on the button and reacting accordingly